requirements.txt
altgraph==0.17.4
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.0
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pefile==2023.2.7
pillow==11.3.0
pluggy==1.6.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
